requirements.txt
confluent-kafka
requests
json5
pyspark
apache-airflow
